# docs/openapi2/v3/openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "contact" : {
      "email" : "contact@email.com",
      "name" : "Contact Name",
      "url" : "http://losol.github.io/eventuras/openapi/v3/eventuras-api.v3.yaml"
    },
    "description" : "Event Management API.",
    "license" : {
      "name" : "GPL"
    },
    "title" : "Eventuras API",
    "version" : "3.0.1"
  },
...
  "servers" : [ {
    "url" : "https://full.path.to/v3/"
  } ],
  "security" : [ {
    "bearer" : [ ]
  } ],
  "paths" : {
...
    "/organizations/{id}" : {
      "delete" : {
        "operationId" : "deleteOrganization",
        "parameters" : [ {
          "description" : "Organization ID.",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Organization deleted."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Access token is missing or invalid."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Access forbidden."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "The requested resource not found."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Unexpected server error."
          }
        },
        "summary" : "Delete organization (system admin only).",
        "tags" : [ "organizations" ]
      },
      "put" : {
        "operationId" : "updateOrganization",
        "parameters" : [ {
          "description" : "Organization ID.",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OrganizationFormDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrganizationDto"
                }
              }
            },
            "description" : "Registration updated."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Access token is missing or invalid."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Access forbidden."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "The requested resource not found."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Unexpected server error."
          }
        },
        "summary" : "Update organization (system admin only).",
        "tags" : [ "organizations" ]
      }
    },
...
  },
  "components" : {
...
    "schemas" : {
      "Error" : {
        "example" : {
          "message" : "Phone and password don't match",
          "code" : "password-mismatch",
          "fields" : {
            "password" : [ "Invalid password" ]
          }
        },
        "properties" : {
          "message" : {
            "description" : "User-friendly error message.",
            "type" : "string"
          },
          "code" : {
            "description" : "Machine-friendly error code.",
            "type" : "string"
          },
          "fields" : {
            "items" : {
              "additionalProperties" : {
                "items" : {
                  "type" : "string"
                },
                "type" : "array"
              },
              "type" : "object"
            },
            "type" : "array"
          }
        },
        "required" : [ "message" ],
        "title" : "Common error response structure.",
        "type" : "object"
      },
...
      "OrganizationDto" : {
        "example" : {
          "name" : "Test org"
        },
        "properties" : {
          "organizationId" : {
            "description" : "Organization ID.",
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "logoUrl" : {
            "type" : "string"
          },
          "løogoBase64" : {
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "title" : "Organization properties.",
        "type" : "object"
      },
      "OrganizationFormDto" : {
        "example" : {
          "name" : "Test org"
        },
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "logoUrl" : {
            "type" : "string"
          },
          "løogoBase64" : {
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "title" : "Organization data.",
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}